+++ src/forest.rs
@@ -1190,5 +1190,5 @@ where
 #[cfg(test)]
 pub(super) mod tests;
 
-#[cfg(any(test, feature = "internal_checks"))]
+#[cfg(any(test, fuzzing))]
 pub mod internal_checks;